+++ kong/plugins/openid-connect/responses.lua
@@ -7,23 +7,13 @@
 
 local uri       = require "kong.openid-connect.uri"
 local log       = require "kong.plugins.openid-connect.log"
+local redirect  = require "kong.plugins.openid-connect.redirect"
 local consumers = require "kong.plugins.openid-connect.consumers"
 
 
 local kong      = kong
 local select    = select
 local concat    = table.concat
-
-
-local REDIRECT_HEADERS = {
-  Location = ""
-}
-
-
-local function redirect(location, status)
-  REDIRECT_HEADERS.Location = location
-  return kong.response.exit(status or 302, "", REDIRECT_HEADERS)
-end
 
 
 local function unauthorized(err, msg, ctx, issuer, client, anonymous, session)