+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix consumer and discovery invalidation events that were returning when the operation
   was `create`. This could leave some cache entries in cache that need to be invalidated.
+- Fix a circular dependency issue with `redirect` function
 
 ## v1.8.1